pyproject.toml
[project]
name = "fine-tuning-bible"
version = "0.1.0"
description = "QLoRA fine-tuning pipeline for Traditional Chinese Bible QA assistant"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.45.0",
    "datasets>=3.0.0",
    "accelerate>=1.0.0",
    "peft>=0.13.0",
    "trl>=0.12.0",
    "bitsandbytes>=0.44.0",
    "trackio>=0.2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "rouge-score>=0.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
